Abstract task choosing into TaskMap component
